End of version 5?
Added more code to the starting of a settlement, but this hasn't been tested yet.

I have started to realize that this project is heading in the wrong direction. This game is a conceptual expansion of DanIdle, to include attackers and the need to defend against them, with a multiplayer aspect. This has become not much like DanIdle at all. A fully new version will be needed to fix that.