Basic map behaviors
We have some basic constructs now to work with, for the game map. The user can scroll around (even if it crashes after releasing b/c a block is selected and no building types are passed in). We also have block types on the left, and a space to describe selected blocks on the right