Completed stick maker
Stick maker is now complete. This requires tools to run, and you start with two options (more comes later, maybe). This also prompted a reorganization so that the number of hooks doesn't change; React was complaining of this. The new code isn't great but I don't know of another solution right now