Toolbox updates
Tool boxes now send a guy to the block requesting tools, when a tool is needed. It's taken a bit of work to get going, so they only head there, never return (that'll be added next). Tool-using blocks will also need to be updated to request tools repeatedly, so that they can actually get a tool. The 'walker' image needs to be reduced in size, too.