Created blockHasSelectableCrafting
Created another add-on component, blockHasSelectableCrafting, for any block that has more than one option for players to craft. This has been applied to the stick maker, but not the flint knapper (it needs to be). Fixed some errors with the toolbox, too. I'll be making blockRequiresTools next, and then we can move forward