Added Hunting Post, Butcher Shop
Added the hunting post and butcher shop... not sure if the butcher shop is functioning yet. Also added blockHasOutputsPerInput, as this functionality will be needed later as well.

Next on the tech tree is the campfire (the butcher shop isn't necessary to cook meats but is helpful, especially for byproducts), which will then allow population to grow. But first we need a way to save (and reload) the local map content.