Fixed Toolbox bugs
I think I've finally got the toolbox working as it should. It seems that multiple tool requests (that blocks are designed to do) generated multiple Traveler objects, each carrying a tool. Since the extra Travelers were never handled, the tools included with them would vanish, and have to be replaced.