Finished Recycler, for now
The Recycler block is now working like it should... for the parts it has. I had planned on allowing players to specify destinations for each part, but I haven't coded that in yet. It can be added later; players can use Haulers to move items out of the block as well.