new leanto
New code has been applied. The Lean-To now works according to the new plans. Currently updating ForagePost to work with new plans; only a few more changes before testing.
Also moved DraggableMap to the libs folder, treating it as a separate entity of sorts. Added comments to it to be more stand-alone. If someone thinks this would be a useful npm module, I'd like to know how to make that happen