Got target tiles to work
When users are assigning tasks, they can now select which tile to perform the task at. We even have tiles highlighted red or green to denote which tiles can be assigned.
However, this only works for one task (collecting twine strips). The code changes need to be applied to the rest of the remote-tile-based tasks