Save-game fixes, expanding tech tree
Made the necessary changes to how games are saved to the server, and everything seems to be working well. Still need to work out loading content too, or else all that work is for nothing.
I also added more notes to the tech tree, and it's kind of exploding in many different directions, and I'm eager to start on it... but gotta get the game loading properly first