worldmap, moved notes
Added more to the world map code. We are now ready to work on sending units; outfitting journeys may be a bit of a challenge to code, but should be interesting. I also moved all the .md files (except Readme) to a notes folder. Added ideas on a world history in a new notes file